refactor(coin-evm): move computeIntentType method to generic alpaca
