+++ .changeset/healthy-hotels-develop.md
@@ -1,0 +1,6 @@
+---
+"@ledgerhq/coin-evm": patch
+"@ledgerhq/live-common": patch
+---
+
+coin-evm: Move computeIntentType method out of AlpacaApi implementation
