--- libs/coin-modules/coin-evm/src/logic/computeIntentType.ts
@@ -1,19 +1,0 @@
-export function computeIntentType(transaction: Record<string, unknown>): string {
-  if (typeof transaction.mode !== "string") {
-    throw new Error(`Unsupported transaction mode: ${transaction.mode}`);
-  }
-
-  if (["send-legacy", "send-eip1559"].includes(transaction.mode)) {
-    return transaction.mode;
-  }
-
-  if (["delegate", "redelegate", "undelegate"].includes(transaction.mode)) {
-    return transaction.type === 2 ? "staking-eip1559" : "staking-legacy";
-  }
-
-  if (transaction.mode === "send") {
-    return transaction.type === 2 ? "send-eip1559" : "send-legacy";
-  }
-
-  throw new Error(`Unsupported transaction mode: '${transaction.mode}'`);
-}
