+++ libs/coin-modules/coin-evm/src/logic/index.ts
@@ -8,7 +8,6 @@ export * from "./craftTransaction";
 export * from "./getNextSequence";
 export * from "./validateIntent";
 export * from "./getStakes";
-export * from "./computeIntentType";
 export * from "./refreshOperations";
 export * from "./getBlock";
 export * from "./getBlockInfo";
