+++ libs/ledger-live-common/src/bridge/generic-alpaca/families/evm/bridge.ts
@@ -33,10 +33,31 @@ export function getAssetFromToken(
   };
 }
 
+export function computeIntentType(transaction: Record<string, unknown>): string {
+  if (typeof transaction.mode !== "string") {
+    throw new Error(`Unsupported transaction mode: ${transaction.mode}`);
+  }
+
+  if (["send-legacy", "send-eip1559"].includes(transaction.mode)) {
+    return transaction.mode;
+  }
+
+  if (["delegate", "redelegate", "undelegate"].includes(transaction.mode)) {
+    return transaction.type === 2 ? "staking-eip1559" : "staking-legacy";
+  }
+
+  if (transaction.mode === "send") {
+    return transaction.type === 2 ? "send-eip1559" : "send-legacy";
+  }
+
+  throw new Error(`Unsupported transaction mode: '${transaction.mode}'`);
+}
+
 export default function evmBridge(currency: CryptoCurrency): BridgeApi {
   return {
     getTokenFromAsset: async (asset: AssetInfo) => getTokenFromAsset(currency, asset),
     getAssetFromToken: (token: TokenCurrency, owner: string) =>
       getAssetFromToken(currency, token, owner),
+    computeIntentType: (transaction: Record<string, unknown>) => computeIntentType(transaction),
   };
 }
